Add files via upload
Regression's Conclusions:
Even if both the page and the country are included in the model, there is still no significant conversion effect.

Results for logistic regression models show that in predicting conversion rate, none of the component, treatment, or country is significant. In order to help predict conversion rate, the e-commerce company should really consider either running the experiment longer or using their domain expertise to define some potential key features of their best customers.

Findings
Refer to Statistical Importance:

Using various ways to view the results, all evidence from the above calculations confirm that there is no need to move to the new page as over all the sample size, we have no evidence to refute the null hypothesis and become more and more confident that the old page is better or at least as good as the new page according to diverted proportion.

Refer to Practical Importance: The change in the diverted rate is 1.04 times what it was before making some change at the most positive diverted rate. So, for designing the new page, we will lose time and money and get marginally positive improvement in diverted rate or even get the worst negative chabge. Often, if the new page is used, the old users of the page will need time to get used to the new page design, which also negatively impacts the diverted rate.